tg_ui_bot/bot_in_python/requirements.txt
setuptools
python-telegram-bot
pillow
requests
